app/requirements.txt
Django==5.2.5
gunicorn
mysqlclient
python-dotenv
djangorestframework
djangorestframework-simplejwt